Add option to end canceling of events

+++ app/controllers/registrations_controller.rb
@@ -54,7 +54,11 @@ class RegistrationsController < ApplicationController
   def destroy_cancel
     registration = Registration.find_by! id: params[:id], barcode: params[:barcode]
     event = registration.event
-    registration.destroy!
+    if registration.event.registration_cancelable
+      registration.destroy!
+    else
+      redirect_to action: :show_cancel
+    end
     RegistrationMailer.confirm_cancel(registration, event).deliver_now
   end
